Make data syncing less sensitive to missing data
Some data may be missing because it is never uploaded, either by mistake on purpose. The missing
data makes the data sync record grow even if the node attempts to keep as few intervals
as possible. It takes longer to share sync records with peers.

The commit makes the syncing process periodically "compact" the older part of the sync record
introducing false positives. The change allows to always keep the record reasonably small.

Additionally, the commit speeds up syncing by not persisting the new sync record on every
chunk write and fixes a bug where the nodes would attempt to sync contiguous intervals of
only half the size intended to keep the number of synced intervals below 10000.